Update header, footer to HeaderLayout, Foorter Layout and divide styles into component for reusable
