+++ src/components/DividerWithText.tsx
@@ -1,6 +1,5 @@
 import React from 'react';
 import { Button, Container } from '@material-ui/core';
-import { yellow } from '@material-ui/core/colors';
 import { makeStyles } from '@material-ui/core/styles';
 
 const useStyles = makeStyles((theme) => ({
