+++ src/components/HeaderLayout.tsx
@@ -1,0 +1,18 @@
+import { FC } from 'react';
+import { AppBar } from '@material-ui/core';
+import Link from 'next/link';
+
+const Header: FC = () => {
+  return (
+    <AppBar color="default">
+      <Link href="/">
+        <a>Home</a>
+      </Link>
+      <Link href="/signup">
+        <a>SignUp</a>
+      </Link>
+    </AppBar>
+  );
+};
+
+export default Header;
