--- src/components/header.tsx
@@ -1,11 +1,0 @@
-import { FC } from 'react';
-
-const Header: FC = () => {
-  return (
-    <header>
-      <nav>nav</nav>
-    </header>
-  );
-};
-
-export default Header;
