+++ src/layouts/AppLayout.tsx
@@ -1,7 +1,12 @@
-import Footer from '../components/Footer';
-import Header from '../components/Header';
+import React from 'react';
+import Footer from '@components/FooterLayout';
+import Header from '@components/HeaderLayout';
 
-const AppLayout = ({ children }) => {
+interface IProps {
+  children: React.ReactNode;
+}
+
+const AppLayout = ({ children }: IProps) => {
   return (
     <>
       <Header />
